+++ src/helpers/store/gettersBuilder.js
@@ -1,0 +1,9 @@
+export default attributes => {
+	const _getters = {};
+	attributes.forEach(attr => {
+		_getters[attr] = state => {
+			return state[attr];
+		};
+	});
+	return _getters;
+};
